# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartAI Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResult {
  success: boolean;
}

export type ConversationMode = typeof ConversationMode[keyof typeof ConversationMode];


export const ConversationMode = {
  general: 'general',
  learning: 'learning',
  career: 'career',
  document: 'document',
  voice: 'voice',
} as const;

export interface Conversation {
  id: number;
  title: string;
  mode: ConversationMode;
  bookmarked: boolean;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
}

export type ConversationInputMode = typeof ConversationInputMode[keyof typeof ConversationInputMode];


export const ConversationInputMode = {
  general: 'general',
  learning: 'learning',
  career: 'career',
  document: 'document',
  voice: 'voice',
} as const;

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
  mode?: ConversationInputMode;
}

export interface ConversationUpdate {
  /** @minLength 1 */
  title?: string;
  bookmarked?: boolean;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
  /** @nullable */
  metadata?: string | null;
}

export type ConversationWithMessagesMode = typeof ConversationWithMessagesMode[keyof typeof ConversationWithMessagesMode];


export const ConversationWithMessagesMode = {
  general: 'general',
  learning: 'learning',
  career: 'career',
  document: 'document',
  voice: 'voice',
} as const;

export interface ConversationWithMessages {
  id: number;
  title: string;
  mode: ConversationWithMessagesMode;
  bookmarked: boolean;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export type ChatMessageInputMode = typeof ChatMessageInputMode[keyof typeof ChatMessageInputMode];


export const ChatMessageInputMode = {
  general: 'general',
  learning: 'learning',
  career: 'career',
  document: 'document',
  voice: 'voice',
} as const;

export type ChatMessageInputLanguage = typeof ChatMessageInputLanguage[keyof typeof ChatMessageInputLanguage];


export const ChatMessageInputLanguage = {
  en: 'en',
  hi: 'hi',
  mr: 'mr',
} as const;

export interface ChatMessageInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  conversationId?: number | null;
  mode?: ChatMessageInputMode;
  model?: string;
  language?: ChatMessageInputLanguage;
}

export interface ChatResponse {
  content: string;
  conversationId: number;
  messageId: number;
  suggestedQuestions: string[];
  model?: string;
}

export interface Document {
  id: number;
  name: string;
  fileType: string;
  fileSize: number;
  createdAt: string;
  analyzed: boolean;
  /** @nullable */
  summary?: string | null;
}

export type DocumentAnalysisRequestAnalysisType = typeof DocumentAnalysisRequestAnalysisType[keyof typeof DocumentAnalysisRequestAnalysisType];


export const DocumentAnalysisRequestAnalysisType = {
  summary: 'summary',
  keyPoints: 'keyPoints',
  qa: 'qa',
  quiz: 'quiz',
} as const;

export interface DocumentAnalysisRequest {
  analysisType: DocumentAnalysisRequestAnalysisType;
}

export interface QAItem {
  question: string;
  answer: string;
}

export interface DocumentAnalysisResult {
  analysisType: string;
  result: string;
  documentId: number;
  qaItems?: QAItem[];
}

export interface Bookmark {
  id: number;
  title: string;
  content: string;
  category: string;
  /** @nullable */
  sourceConversationId?: number | null;
  createdAt: string;
}

export interface BookmarkInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  category: string;
  /** @nullable */
  sourceConversationId?: number | null;
}

export type UserProfilePreferredLanguage = typeof UserProfilePreferredLanguage[keyof typeof UserProfilePreferredLanguage];


export const UserProfilePreferredLanguage = {
  en: 'en',
  hi: 'hi',
  mr: 'mr',
} as const;

export type UserProfileTheme = typeof UserProfileTheme[keyof typeof UserProfileTheme];


export const UserProfileTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface UserProfile {
  id: string;
  /** @nullable */
  displayName: string | null;
  preferredLanguage: UserProfilePreferredLanguage;
  theme: UserProfileTheme;
  preferredModel: string;
  voiceEnabled: boolean;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
}

export type UserProfileUpdatePreferredLanguage = typeof UserProfileUpdatePreferredLanguage[keyof typeof UserProfileUpdatePreferredLanguage];


export const UserProfileUpdatePreferredLanguage = {
  en: 'en',
  hi: 'hi',
  mr: 'mr',
} as const;

export type UserProfileUpdateTheme = typeof UserProfileUpdateTheme[keyof typeof UserProfileUpdateTheme];


export const UserProfileUpdateTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface UserProfileUpdate {
  displayName?: string;
  preferredLanguage?: UserProfileUpdatePreferredLanguage;
  theme?: UserProfileUpdateTheme;
  preferredModel?: string;
  voiceEnabled?: boolean;
  bio?: string;
}

export interface ModeCount {
  mode: string;
  count: number;
}

export interface RecentActivity {
  type: string;
  title: string;
  createdAt: string;
}

export interface DashboardStats {
  totalConversations: number;
  totalMessages: number;
  totalDocuments: number;
  bookmarkedConversations: number;
  conversationsByMode: ModeCount[];
  recentActivity: RecentActivity[];
}

export interface OllamaModel {
  name: string;
  /** @nullable */
  size: number | null;
  /** @nullable */
  digest?: string | null;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};