# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartAI Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List conversations for current user
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [chat]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateConversation
      tags: [chat]
      summary: Update conversation title or bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationUpdate"
      responses:
        "200":
          description: Updated conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
    delete:
      operationId: deleteConversation
      tags: [chat]
      summary: Delete a conversation and its messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "503":
          description: Ollama unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: List bookmarks for current user
      responses:
        "200":
          description: List of bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Create a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Bookmark created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Delete a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    Conversation:
      type: object
      required: [id, title, mode, bookmarked, createdAt, updatedAt, messageCount]
      properties:
        id:
          type: integer
        title:
          type: string
        mode:
          type: string
          enum: [general, learning, career, document, voice]
        bookmarked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messageCount:
          type: integer
        lastMessage:
          type: ["string", "null"]
    ConversationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        mode:
          type: string
          enum: [general, learning, career, document, voice]
    ConversationUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        bookmarked:
          type: boolean
    Message:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        metadata:
          type: ["string", "null"]
    ConversationWithMessages:
      type: object
      required: [id, title, mode, bookmarked, createdAt, updatedAt, messages]
      properties:
        id:
          type: integer
        title:
          type: string
        mode:
          type: string
          enum: [general, learning, career, document, voice]
        bookmarked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        conversationId:
          type: ["integer", "null"]
        mode:
          type: string
          enum: [general, learning, career, document, voice]
        model:
          type: string
        language:
          type: string
          enum: [en, hi, mr]
    ChatResponse:
      type: object
      required: [content, conversationId, messageId, suggestedQuestions]
      properties:
        content:
          type: string
        conversationId:
          type: integer
        messageId:
          type: integer
        suggestedQuestions:
          type: array
          items:
            type: string
        model:
          type: string
...
    Bookmark:
      type: object
      required: [id, title, content, category, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        sourceConversationId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    BookmarkInput:
      type: object
      required: [title, content, category]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        category:
          type: string
        sourceConversationId:
          type: ["integer", "null"]
...