# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartAI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Bookmark {
  id: number;
  title: string;
  content: string;
  category: string;
  /** @nullable */
  sourceConversationId?: number | null;
  createdAt: Date;
}