# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartAI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageInputLanguage } from './chatMessageInputLanguage';
import type { ChatMessageInputMode } from './chatMessageInputMode';

export interface ChatMessageInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  conversationId?: number | null;
  mode?: ChatMessageInputMode;
  model?: string;
  language?: ChatMessageInputLanguage;
}