# lib/api-zod/src/generated/types/conversationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartAI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationInputMode } from './conversationInputMode';

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
  mode?: ConversationInputMode;
}